# src/main/java/com/redhat/ecosystemappeng/exploitiq/model/FeedbackDto.java
/*
 * SPDX-FileCopyrightText: Copyright (c) 2026, Red Hat Inc. & AFFILIATES. All rights reserved.
 * SPDX-License-Identifier: Apache-2.0
 * Licensed under the Apache License, Version 2.0 (the "License");
 * you may not use this file except in compliance with the License.
 * You may obtain a copy of the License at
 * http://www.apache.org/licenses/LICENSE-2.0
 * Unless required by applicable law or agreed to in writing, software
 * distributed under the License is distributed on an "AS IS" BASIS,
 * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 * See the License for the specific language governing permissions and
 * limitations under the License.
 */

package com.redhat.ecosystemappeng.exploitiq.model;

import io.quarkus.runtime.annotations.RegisterForReflection;
import jakarta.validation.constraints.Max;
import jakarta.validation.constraints.Min;
import jakarta.validation.constraints.NotBlank;
import jakarta.validation.constraints.NotNull;
import jakarta.validation.constraints.Pattern;
import org.eclipse.microprofile.openapi.annotations.media.Schema;

@RegisterForReflection
@Schema(name = "Feedback", description = "Data Transfer Object for submitting user feedback about a specific report.")
public class FeedbackDto {

    static final String ACCURACY_PATTERN =
            "Very Accurate|Mostly Accurate|Somewhat Inaccurate|Incorrect";
    static final String REASONING_OR_CHECKLIST_PATTERN = "Yes|Mostly|Somewhat|No";

    @NotNull
    @Min(1)
    @Max(5)
    @Schema(required = true, description = "Rating value from 1 to 5", minimum = "1", maximum = "5")
    private Integer rating;

    @Schema(description = "Additional comment or feedback text")
    private String comment;

    @NotBlank
    @Pattern(regexp = ACCURACY_PATTERN, message = "Invalid accuracy value")
    @Schema(
            required = true,
            description = "Accuracy assessment of the report",
            enumeration = {"Very Accurate", "Mostly Accurate", "Somewhat Inaccurate", "Incorrect"})
    private String accuracy;

    @NotBlank
    @Pattern(regexp = REASONING_OR_CHECKLIST_PATTERN, message = "Invalid reasoning value")
    @Schema(
            required = true,
            description = "Reasoning or explanation of the report",
            enumeration = {"Yes", "Mostly", "Somewhat", "No"})
    private String reasoning;

    @NotBlank
    @Pattern(regexp = REASONING_OR_CHECKLIST_PATTERN, message = "Invalid checklist value")
    @Schema(
            required = true,
            description = "Checklist assessment of the report",
            enumeration = {"Yes", "Mostly", "Somewhat", "No"})
    private String checklist;

    public Integer getRating() { return rating; }
    public void setRating(Integer rating) { this.rating = rating; }
    public String getComment() { return comment; }
    public void setComment(String comment) { this.comment = comment; }
    public String getAccuracy() {
        return accuracy;
    }
    public void setAccuracy(String accuracy) {
        this.accuracy = accuracy;
    }
    public String getReasoning() {
        return reasoning;
    }
    public void setReasoning(String reasoning) {
        this.reasoning = reasoning;
    }
    public String getChecklist() {
        return checklist;
    }
    public void setChecklist(String checklist) {
        this.checklist = checklist;
    }
}