# src/main/java/com/redhat/ecosystemappeng/exploitiq/rest/FeedbackResource.java
/*
 * SPDX-FileCopyrightText: Copyright (c) 2026, Red Hat Inc. & AFFILIATES. All rights reserved.
 * SPDX-License-Identifier: Apache-2.0
 * Licensed under the Apache License, Version 2.0 (the "License");
 * you may not use this file except in compliance with the License.
 * You may obtain a copy of the License at
 * http://www.apache.org/licenses/LICENSE-2.0
 * Unless required by applicable law or agreed to in writing, software
 * distributed under the License is distributed on an "AS IS" BASIS,
 * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 * See the License for the specific language governing permissions and
 * limitations under the License.
 */

package com.redhat.ecosystemappeng.exploitiq.rest;

import com.redhat.ecosystemappeng.exploitiq.exception.FeedbackAlreadyExistsException;
import com.redhat.ecosystemappeng.exploitiq.model.ErrorResponse;
import com.redhat.ecosystemappeng.exploitiq.model.FeedbackDto;
import com.redhat.ecosystemappeng.exploitiq.model.FeedbackResponseDto;
import com.redhat.ecosystemappeng.exploitiq.service.FeedbackService;
import jakarta.inject.Inject;
import jakarta.validation.ConstraintViolation;
import jakarta.validation.ConstraintViolationException;
import jakarta.validation.Valid;
import jakarta.ws.rs.*;
import jakarta.ws.rs.NotFoundException;
import jakarta.ws.rs.core.Context;
import jakarta.ws.rs.core.MediaType;
import jakarta.ws.rs.core.Response;
import jakarta.ws.rs.core.SecurityContext;
import org.jboss.logging.Logger;

import org.eclipse.microprofile.openapi.annotations.Operation;
import org.eclipse.microprofile.openapi.annotations.media.Content;
import org.eclipse.microprofile.openapi.annotations.media.Schema;
import org.eclipse.microprofile.openapi.annotations.parameters.RequestBody;
import org.eclipse.microprofile.openapi.annotations.parameters.Parameter;
import org.eclipse.microprofile.openapi.annotations.responses.APIResponse;
import org.eclipse.microprofile.openapi.annotations.responses.APIResponses;
import org.jboss.resteasy.reactive.server.ServerExceptionMapper;

import java.util.Objects;

@Path("/reports/{reportId}/feedback")
@Consumes(MediaType.APPLICATION_JSON)
@Produces(MediaType.APPLICATION_JSON)
public class FeedbackResource {
    private static final Logger LOGGER = Logger.getLogger(FeedbackResource.class);

    @Inject
    FeedbackService feedbackService;

    @Context
    SecurityContext securityContext;

    @GET
    @Operation(
      summary = "Get submitted feedback for a report",
      description = "Returns the authenticated user's submitted feedback for the given report, or 404 if none")
    @APIResponses({
      @APIResponse(
        responseCode = "200",
        description = "Feedback retrieved successfully",
        content = @Content(schema = @Schema(implementation = FeedbackResponseDto.class))
      ),
      @APIResponse(
        responseCode = "404",
        description = "No feedback found for this report"
      ),
      @APIResponse(
        responseCode = "500",
        description = "Internal server error"
      )
    })
    public Response getFeedback(
      @Parameter(description = "Report identifier", required = true)
      @PathParam("reportId") String reportId) {
        LOGGER.infof("Fetching feedback for reportId: %s", reportId);
        try {
            FeedbackResponseDto feedback = feedbackService.getFeedback(reportId, securityContext);
            if (Objects.isNull(feedback)) {
                return Response.status(Response.Status.NOT_FOUND)
                        .entity(new ErrorResponse("Feedback not found."))
                        .build();
            }
            return Response.ok(feedback).build();
        } catch (Exception e) {
            LOGGER.error("Failed to fetch feedback:", e);
            return Response.status(Response.Status.INTERNAL_SERVER_ERROR)
                    .entity(new ErrorResponse("Failed to fetch feedback."))
                    .build();
        }
    }

    @POST
    @Operation(
      summary = "Submit user feedback for a report",
      description = "Submits user feedback for the given report and persists it in MongoDB")
    @APIResponses({
      @APIResponse(
        responseCode = "201",
        description = "Feedback successfully created",
        content = @Content(schema = @Schema(implementation = FeedbackResponseDto.class))
      ),
      @APIResponse(
        responseCode = "404",
        description = "Report not found"
      ),
      @APIResponse(
        responseCode = "409",
        description = "Feedback already submitted for this report"
      ),
      @APIResponse(
        responseCode = "400",
        description = "Invalid request body"
      ),
      @APIResponse(
        responseCode = "500",
        description = "Internal server error"
      )
    })
    public Response submitFeedback(
      @Parameter(description = "Report identifier", required = true)
      @PathParam("reportId") String reportId,
      @RequestBody(
        description = "User feedback data",
        required = true,
        content = @Content(schema = @Schema(implementation = FeedbackDto.class))
      )
      @Valid FeedbackDto dto) {
        LOGGER.infof(
                "Received feedback for reportId=[%s], rating=[%d], accuracy=[%s], reasoning=[%s], checklist=[%s]",
                reportId,
                dto.getRating(),
                dto.getAccuracy(),
                dto.getReasoning(),
                dto.getChecklist()
        );
        FeedbackResponseDto saved = feedbackService.sendFeedback(reportId, dto, securityContext);
        return Response.status(Response.Status.CREATED).entity(saved).build();
    }

    @ServerExceptionMapper
    public Response mapAlreadyExistsException(FeedbackAlreadyExistsException e) {
        return Response.status(Response.Status.CONFLICT)
                .entity(new ErrorResponse(e.getMessage()))
                .build();
    }

    @ServerExceptionMapper
    public Response mapNotFoundException(NotFoundException e) {
        return Response.status(Response.Status.NOT_FOUND)
                .entity(new ErrorResponse(e.getMessage()))
                .build();
    }

    @ServerExceptionMapper
    public Response mapConstraintViolationException(ConstraintViolationException e) {
        String message = e.getConstraintViolations().stream()
                .map(ConstraintViolation::getMessage)
                .findFirst()
                .orElse("Invalid request");
        return Response.status(Response.Status.BAD_REQUEST)
                .entity(new ErrorResponse(message))
                .build();
    }

    @ServerExceptionMapper
    public Response mapException(Exception e) {
        LOGGER.error("Unexpected error in FeedbackResource", e);
        return Response.status(Response.Status.INTERNAL_SERVER_ERROR)
                .entity(new ErrorResponse("An unexpected error occurred"))
                .build();
    }

}