# src/main/webui/openapi.json
{
  "openapi": "3.1.0",
  "components": {
    "schemas": {
...
      "Feedback": {
        "description": "Data Transfer Object for submitting user feedback about a specific report.",
        "type": "object",
        "required": [
          "rating",
          "accuracy",
          "reasoning",
          "checklist"
        ],
        "properties": {
          "rating": {
            "type": "integer",
            "format": "int32",
            "maximum": 5,
            "minimum": 1,
            "description": "Rating value from 1 to 5"
          },
          "comment": {
            "type": "string",
            "description": "Additional comment or feedback text"
          },
          "accuracy": {
            "type": "string",
            "description": "Accuracy assessment of the report"
          },
          "reasoning": {
            "type": "string",
            "description": "Reasoning or explanation of the report"
          },
          "checklist": {
            "type": "string",
            "description": "Checklist assessment of the report"
          }
        }
      },
      "FeedbackResponse": {
        "description": "Submitted feedback for a report.",
        "type": "object",
        "required": [
          "rating",
          "accuracy",
          "reasoning",
          "checklist",
          "submittedAt"
        ],
        "properties": {
          "rating": {
            "type": "integer",
            "format": "int32",
            "maximum": 5,
            "minimum": 1,
            "description": "Rating value from 1 to 5"
          },
          "comment": {
            "type": "string",
            "description": "Additional comment or feedback text"
          },
          "accuracy": {
            "type": "string",
            "description": "Accuracy assessment of the report"
          },
          "reasoning": {
            "type": "string",
            "description": "Reasoning or explanation of the report"
          },
          "checklist": {
            "type": "string",
            "description": "Checklist assessment of the report"
          },
          "submittedAt": {
            "type": "string",
            "description": "Timestamp when feedback was submitted"
          }
        }
      },
...
    },
    "securitySchemes": {
      "jwt": {
        "type": "http",
        "description": "Please enter your JWT Token without Bearer",
        "scheme": "bearer",
        "bearerFormat": "jwt"
      }
    }
  },
...
  "paths": {
...
    "/api/v1/reports/{reportId}/feedback": {
      "get": {
        "summary": "Get submitted feedback for a report",
        "description": "Returns the authenticated user's submitted feedback for the given report, or 404 if none",
        "parameters": [
          {
            "description": "Report identifier",
            "required": true,
            "name": "reportId",
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Feedback retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FeedbackResponse"
                }
              }
            }
          },
          "404": {
            "description": "No feedback found for this report"
          },
          "500": {
            "description": "Internal server error"
          }
        },
        "tags": [
          "Feedback Resource"
        ]
      },
      "post": {
        "summary": "Submit user feedback for a report",
        "description": "Submits user feedback for the given report and persists it in MongoDB",
        "parameters": [
          {
            "description": "Report identifier",
            "required": true,
            "name": "reportId",
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "description": "User feedback data",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Feedback"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Feedback successfully created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FeedbackResponse"
                }
              }
            }
          },
          "409": {
            "description": "Feedback already submitted for this report"
          },
          "400": {
            "description": "Invalid request body"
          },
          "404": {
            "description": "Report not found"
          },
          "500": {
            "description": "Internal server error"
          }
        },
        "tags": [
          "Feedback Resource"
        ]
      }
    },
...
  },
  "info": {
    "title": "exploit-iq-client API",
    "version": "1.0.0-SNAPSHOT"
  }
}