# src/main/webui/src/generated-client/models/FeedbackResponse.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
/**
 * Submitted feedback for a report.
 */
export type FeedbackResponse = {
    rating: number;
    comment?: string;
    accuracy: string;
    reasoning: string;
    checklist: string;
    submittedAt: string;
};